+++ CHANGE_LOG.md
@@ -79,3 +79,10 @@
   + Remove command to start trackpoint.sh in Option 4 (redundant since settings already applied in Option 2)
 - v1.7.3
   + Add initially tested OS/Laptops to README
+##### v1.8.0
+    + Add check for root user
+    + Add validation of setting values
+    + Prevent using values entered if any are not valid
+    + Display current corresponding trackpoint setting when setting new values
+    + Add better detection of trackpoint existence
+    + Initialize trackpoint variables with current setting values